feat(pkg/test): export NoWrapper render option
