+++ extension/extension_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/bevzzz/nb"
 	"github.com/bevzzz/nb/extension"
 	"github.com/bevzzz/nb/pkg/test"
-	"github.com/bevzzz/nb/render"
 	"github.com/bevzzz/nb/schema"
 	"github.com/stretchr/testify/require"
 )
@@ -17,16 +16,16 @@ func TestMarkdown(t *testing.T) {
 	// Arrange
 	var sb strings.Builder
 	want := "Hi, mom!"
-	c := nb.New(nb.WithExtensions(
-		extension.NewMarkdown(func(w io.Writer, c schema.Cell) error {
-			io.WriteString(w, want)
-			return nil
-		}),
-	))
-
-	// Override default CellWrapper to compare bare cell contents only.
+	c := nb.New(
+		nb.WithExtensions(
+			extension.NewMarkdown(func(w io.Writer, c schema.Cell) error {
+				io.WriteString(w, want)
+				return nil
+			}),
+		),
+		nb.WithRenderOptions(test.NoWrapper),
+	)
 	r := c.Renderer()
-	r.AddOptions(render.WithCellRenderers(&fakeWrapper{}))
 
 	// Act
 	err := r.Render(&sb, test.Notebook(test.Markdown("Bye!")))
@@ -60,16 +59,16 @@ func TestStream(t *testing.T) {
 			// Arrange
 			var sb strings.Builder
 			want := "Hi, mom!"
-			c := nb.New(nb.WithExtensions(
-				extension.NewStream(func(w io.Writer, c schema.Cell) error {
-					io.WriteString(w, want)
-					return nil
-				}),
-			))
-
-			// Override default CellWrapper to compare bare cell contents only.
+			c := nb.New(
+				nb.WithExtensions(
+					extension.NewStream(func(w io.Writer, c schema.Cell) error {
+						io.WriteString(w, want)
+						return nil
+					}),
+				),
+				nb.WithRenderOptions(test.NoWrapper),
+			)
 			r := c.Renderer()
-			r.AddOptions(render.WithCellRenderers(&fakeWrapper{}))
 
 			// Act
 			err := r.Render(&sb, test.Notebook(tt.cell))
@@ -81,23 +80,4 @@ func TestStream(t *testing.T) {
 			}
 		})
 	}
-}
-
-// fakeWrapper calls the passed RenderCellFunc immediately without any additional writes to w.
-type fakeWrapper struct{}
-
-var _ render.CellWrapper = (*fakeWrapper)(nil)
-
-func (*fakeWrapper) RegisterFuncs(render.RenderCellFuncRegistry)                    {}
-func (*fakeWrapper) Wrap(w io.Writer, c schema.Cell, r render.RenderCellFunc) error { return r(w, c) }
-func (*fakeWrapper) WrapInput(w io.Writer, c schema.Cell, r render.RenderCellFunc) error {
-	return r(w, c)
-}
-func (*fakeWrapper) WrapOutput(w io.Writer, out schema.Outputter, r render.RenderCellFunc) error {
-	for _, c := range out.Outputs() {
-		if err := r(w, c); err != nil {
-			return err
-		}
-	}
-	return nil
 }
